refactor: extract live device hook
